requirements.txt
# --- Google API Core ---
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# --- Gmail API ---
google-api-python-client==2.148.0
googleapis-common-protos==1.64.0

# --- Google Sheets API ---
gspread==6.1.2

# --- Data utils ---
pandas==2.2.3
python-dateutil==2.9.0.post0

# --- Email parsing ---
email-validator==2.2.0
beautifulsoup4==4.12.3
lxml==5.3.0
# Note: bs4 is just an alias for beautifulsoup4, not needed separately

# --- Optional ---
redis==5.0.7
aiohttp==3.10.10

# --- Auth / dotenv / logging ---
python-dotenv==1.0.1
loguru==0.7.2

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0